requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace ecosystem
transformers>=4.35.0
accelerate>=0.24.0

# Dataset / dataframe handling
pandas>=2.0.0
Pillow>=9.0.0

# Metrics
scikit-learn>=1.3.0
nltk>=3.8.0

# Visualisation
matplotlib>=3.7.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
